third_party/blink/web_tests/external/wpt/html/dom/reflection-metadata-expected.txt

This is a testharness.js-based test.
Found 86 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] link.as: setAttribute() to "audio"
  assert_equals: IDL get expected "audio" but got ""
[FAIL] link.as: setAttribute() to "AUDIO"
  assert_equals: IDL get expected "audio" but got ""
[FAIL] link.as: setAttribute() to "document"
  assert_equals: IDL get expected "document" but got ""
[FAIL] link.as: setAttribute() to "DOCUMENT"
  assert_equals: IDL get expected "document" but got ""
[FAIL] link.as: setAttribute() to "embed"
  assert_equals: IDL get expected "embed" but got ""
[FAIL] link.as: setAttribute() to "EMBED"
  assert_equals: IDL get expected "embed" but got ""
[FAIL] link.as: setAttribute() to "manifest"
  assert_equals: IDL get expected "manifest" but got ""
[FAIL] link.as: setAttribute() to "MANIFEST"
  assert_equals: IDL get expected "manifest" but got ""
[FAIL] link.as: setAttribute() to "object"
  assert_equals: IDL get expected "object" but got ""
[FAIL] link.as: setAttribute() to "OBJECT"
  assert_equals: IDL get expected "object" but got ""
[FAIL] link.as: setAttribute() to "report"
  assert_equals: IDL get expected "report" but got ""
[FAIL] link.as: setAttribute() to "REPORT"
  assert_equals: IDL get expected "report" but got ""
[FAIL] link.as: setAttribute() to "sharedworker"
  assert_equals: IDL get expected "sharedworker" but got ""
[FAIL] link.as: setAttribute() to "SHAREDWORKER"
  assert_equals: IDL get expected "sharedworker" but got ""
[FAIL] link.as: setAttribute() to "video"
  assert_equals: IDL get expected "video" but got ""
[FAIL] link.as: setAttribute() to "VIDEO"
  assert_equals: IDL get expected "video" but got ""
[FAIL] link.as: setAttribute() to "worker"
  assert_equals: IDL get expected "worker" but got ""
[FAIL] link.as: setAttribute() to "WORKER"
  assert_equals: IDL get expected "worker" but got ""
[FAIL] link.as: setAttribute() to "xslt"
  assert_equals: IDL get expected "xslt" but got ""
[FAIL] link.as: setAttribute() to "XSLT"
  assert_equals: IDL get expected "xslt" but got ""
[FAIL] link.as: IDL set to "audio"
  assert_equals: IDL get expected "audio" but got ""
[FAIL] link.as: IDL set to "AUDIO"
  assert_equals: IDL get expected "audio" but got ""
[FAIL] link.as: IDL set to "document"
  assert_equals: IDL get expected "document" but got ""
[FAIL] link.as: IDL set to "DOCUMENT"
  assert_equals: IDL get expected "document" but got ""
[FAIL] link.as: IDL set to "embed"
  assert_equals: IDL get expected "embed" but got ""
[FAIL] link.as: IDL set to "EMBED"
  assert_equals: IDL get expected "embed" but got ""
[FAIL] link.as: IDL set to "manifest"
  assert_equals: IDL get expected "manifest" but got ""
[FAIL] link.as: IDL set to "MANIFEST"
  assert_equals: IDL get expected "manifest" but got ""
[FAIL] link.as: IDL set to "object"
  assert_equals: IDL get expected "object" but got ""
[FAIL] link.as: IDL set to "OBJECT"
  assert_equals: IDL get expected "object" but got ""
[FAIL] link.as: IDL set to "report"
  assert_equals: IDL get expected "report" but got ""
[FAIL] link.as: IDL set to "REPORT"
  assert_equals: IDL get expected "report" but got ""
[FAIL] link.as: IDL set to "sharedworker"
  assert_equals: IDL get expected "sharedworker" but got ""
[FAIL] link.as: IDL set to "SHAREDWORKER"
  assert_equals: IDL get expected "sharedworker" but got ""
[FAIL] link.as: IDL set to "video"
  assert_equals: IDL get expected "video" but got ""
[FAIL] link.as: IDL set to "VIDEO"
  assert_equals: IDL get expected "video" but got ""
[FAIL] link.as: IDL set to "worker"
  assert_equals: IDL get expected "worker" but got ""
[FAIL] link.as: IDL set to "WORKER"
  assert_equals: IDL get expected "worker" but got ""
[FAIL] link.as: IDL set to "xslt"
  assert_equals: IDL get expected "xslt" but got ""
[FAIL] link.as: IDL set to "XSLT"
  assert_equals: IDL get expected "xslt" but got ""
[FAIL] link.nonce: IDL set to ""
  assert_equals: getAttribute() expected "" but got "test-valueOf"
[FAIL] link.nonce: IDL set to " \\0\\x01\\x02\\x03\\x04\\x05\\x06\\x07 \\b\\t\\n\\v\\f\\r\\x0e\\x0f \\x10\\x11\\x12\\x13\\x14\\x15\\x16\\x17 \\x18\\x19\\x1a\\x1b\\x1c\\x1d\\x1e\\x1f  foo "
  assert_equals: getAttribute() expected " \\0\\x01\\x02\\x03\\x04\\x05\\x06\\x07 \\b\\t\\n\\v\\f\\r\\x0e\\x0f \\x10\\x11\\x12\\x13\\x14\\x15\\x16\\x17 \\x18\\x19\\x1a\\x1b\\x1c\\x1d\\x1e\\x1f  foo " but got "test-valueOf"
[FAIL] link.nonce: IDL set to undefined
  assert_equals: getAttribute() expected "undefined" but got "test-valueOf"
[FAIL] link.nonce: IDL set to 7
  assert_equals: getAttribute() expected "7" but got "test-valueOf"
[FAIL] link.nonce: IDL set to 1.5
  assert_equals: getAttribute() expected "1.5" but got "test-valueOf"
[FAIL] link.nonce: IDL set to "5%"
  assert_equals: getAttribute() expected "5%" but got "test-valueOf"
[FAIL] link.nonce: IDL set to "+100"
  assert_equals: getAttribute() expected "+100" but got "test-valueOf"
[FAIL] link.nonce: IDL set to ".5"
  assert_equals: getAttribute() expected ".5" but got "test-valueOf"
[FAIL] link.nonce: IDL set to true
  assert_equals: getAttribute() expected "true" but got "test-valueOf"
[FAIL] link.nonce: IDL set to false
  assert_equals: getAttribute() expected "false" but got "test-valueOf"
[FAIL] link.nonce: IDL set to object "[object Object]"
  assert_equals: getAttribute() expected "[object Object]" but got "test-valueOf"
[FAIL] link.nonce: IDL set to NaN
  assert_equals: getAttribute() expected "NaN" but got "test-valueOf"
[FAIL] link.nonce: IDL set to Infinity
  assert_equals: getAttribute() expected "Infinity" but got "test-valueOf"
[FAIL] link.nonce: IDL set to -Infinity
  assert_equals: getAttribute() expected "-Infinity" but got "test-valueOf"
[FAIL] link.nonce: IDL set to "\\0"
  assert_equals: getAttribute() expected "\\0" but got "test-valueOf"
[FAIL] link.nonce: IDL set to null
  assert_equals: getAttribute() expected "null" but got "test-valueOf"
[FAIL] link.nonce: IDL set to object "test-toString"
  assert_equals: getAttribute() expected "test-toString" but got "test-valueOf"
[FAIL] link.referrerPolicy: setAttribute() to "same-origin"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] link.referrerPolicy: setAttribute() to "SAME-ORIGIN"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] link.referrerPolicy: setAttribute() to "strict-origin"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] link.referrerPolicy: setAttribute() to "STRICT-ORIGIN"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] link.referrerPolicy: setAttribute() to "strict-origin-when-cross-origin"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] link.referrerPolicy: setAttribute() to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] link.referrerPolicy: IDL set to "same-origin"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] link.referrerPolicy: IDL set to "SAME-ORIGIN"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] link.referrerPolicy: IDL set to "strict-origin"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] link.referrerPolicy: IDL set to "STRICT-ORIGIN"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] link.referrerPolicy: IDL set to "strict-origin-when-cross-origin"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] link.referrerPolicy: IDL set to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] style.nonce: IDL set to ""
  assert_equals: getAttribute() expected "" but got "test-valueOf"
[FAIL] style.nonce: IDL set to " \\0\\x01\\x02\\x03\\x04\\x05\\x06\\x07 \\b\\t\\n\\v\\f\\r\\x0e\\x0f \\x10\\x11\\x12\\x13\\x14\\x15\\x16\\x17 \\x18\\x19\\x1a\\x1b\\x1c\\x1d\\x1e\\x1f  foo "
  assert_equals: getAttribute() expected " \\0\\x01\\x02\\x03\\x04\\x05\\x06\\x07 \\b\\t\\n\\v\\f\\r\\x0e\\x0f \\x10\\x11\\x12\\x13\\x14\\x15\\x16\\x17 \\x18\\x19\\x1a\\x1b\\x1c\\x1d\\x1e\\x1f  foo " but got "test-valueOf"
[FAIL] style.nonce: IDL set to undefined
  assert_equals: getAttribute() expected "undefined" but got "test-valueOf"
[FAIL] style.nonce: IDL set to 7
  assert_equals: getAttribute() expected "7" but got "test-valueOf"
[FAIL] style.nonce: IDL set to 1.5
  assert_equals: getAttribute() expected "1.5" but got "test-valueOf"
[FAIL] style.nonce: IDL set to "5%"
  assert_equals: getAttribute() expected "5%" but got "test-valueOf"
[FAIL] style.nonce: IDL set to "+100"
  assert_equals: getAttribute() expected "+100" but got "test-valueOf"
[FAIL] style.nonce: IDL set to ".5"
  assert_equals: getAttribute() expected ".5" but got "test-valueOf"
[FAIL] style.nonce: IDL set to true
  assert_equals: getAttribute() expected "true" but got "test-valueOf"
[FAIL] style.nonce: IDL set to false
  assert_equals: getAttribute() expected "false" but got "test-valueOf"
[FAIL] style.nonce: IDL set to object "[object Object]"
  assert_equals: getAttribute() expected "[object Object]" but got "test-valueOf"
[FAIL] style.nonce: IDL set to NaN
  assert_equals: getAttribute() expected "NaN" but got "test-valueOf"
[FAIL] style.nonce: IDL set to Infinity
  assert_equals: getAttribute() expected "Infinity" but got "test-valueOf"
[FAIL] style.nonce: IDL set to -Infinity
  assert_equals: getAttribute() expected "-Infinity" but got "test-valueOf"
[FAIL] style.nonce: IDL set to "\\0"
  assert_equals: getAttribute() expected "\\0" but got "test-valueOf"
[FAIL] style.nonce: IDL set to null
  assert_equals: getAttribute() expected "null" but got "test-valueOf"
[FAIL] style.nonce: IDL set to object "test-toString"
  assert_equals: getAttribute() expected "test-toString" but got "test-valueOf"
Harness: the test ran to completion.
